Check for custom emoji for message length and add leniency
Change from message.content to message.clean_content, this makes sure channel mentions etc. get counted by the channel name, not the channel ID (which will always be 17-19, even for channels just called "faq" for example.
The regex I added checks for custom animated and regular emoji (hence the a?). Early IDs (2015) will be 17 numbers, older IDs will be 18 or 19 in length.
We give a flat 20 int amount of leniency since checking for the exact amount is overkill.
The added length added because we get the raw content is between 20-23 chars, so just subtracting 20 should be fine. This will make emoji spam more possible, which is why another check for 4+ emoji or similar with the length check would be sensible but I didn't add it since this is already a PR w/o an Issue and I don't want to add more assumptions to how you want to run this.